refactor(sharepoint): share Graph transport via core/msgraph; add docs
Extract the duplicated Microsoft Graph transport into a neutral
core/msgraph module (config + client), mirroring core/google. SharePoint
now re-exports the shared transport and keeps only its own drive-id
addressing. OneDrive is left untouched.

Add SharePoint setup and Python resource docs.
